src/test/java/com/shonlee/service/GirlServiceTest.java
package com.shonlee.service;

import com.shonlee.domain.Girl;
import org.junit.Assert;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;

import static org.junit.Assert.*;

/**
 * Created by ShonLee on 2017/6/10.
 */
@RunWith(SpringRunner.class)
@SpringBootTest
public class GirlServiceTest {
    @Autowired
    private GirlService girlService;
    @Test
    public void findOne() throws Exception {
        Girl girl=girlService.findOne(13);
        Assert.assertEquals(new Integer(28),girl.getAge());
    }
}